+++ game/Scripts/Framework/MainScene.gd
@@ -1,11 +1,12 @@
 extends Node
 
 export (Array, PackedScene) var PanelSceneList
-var panelMgr
+
 var logMgr
+var panelMgr
 
 func _ready():
+	logMgr = get_node("/root/LogMgr")
 	panelMgr = get_node("/root/PanelMgr")
-	logMgr = get_node("/root/GLog")
 	logMgr._log("[MainScene] ready")
 	panelMgr.setMainScene(self)