+++ game/Scripts/Math/MathScoreData.gd
@@ -1,9 +1,5 @@
-extends Control
+extends PanelBasic
 
-var dataMgr: GData
-var logMgr: GLog
-var panelMgr: PanelMgr
-var scoreMgr: GScore
 
 onready var ppLabelSummary: RichTextLabel = $VBoxContainer/MarginContainer/Swipe/HBoxContainer/ScrollContainer/VBoxContainer/LabelSummary
 onready var ppLabelDatetime: RichTextLabel = $VBoxContainer/MarginContainer/Swipe/HBoxContainer/LabelDatetime
@@ -11,26 +7,15 @@ onready var ppLabelList: RichTextLabel = $VBoxContainer/MarginContainer/Swipe/HB
 onready var ppMap: TextureRect = $VBoxContainer/MarginContainer/Swipe/HBoxContainer/ScrollContainer/VBoxContainer/HBoxContainer/MathMulMap
 onready var ppSwipe: PanelSwipe = $VBoxContainer/MarginContainer/Swipe
 
-func _ready():
-	panelMgr = get_node("/root/PanelMgr")
-	logMgr = get_node("/root/GLog")
-	dataMgr = get_node("/root/GData")
-	scoreMgr = get_node("/root/GScore")
 
-
-func _setRectSize(ssize: Vector2):
-	rect_position = Vector2(0, 0)
-	rect_size = ssize
-	$BasicBg.rect_min_size = ssize
-	$VBoxContainer.rect_min_size = ssize
-	
-	dataMgr._clear_hightlight_color()
+func _panel_ready():
+	_g.dataMgr._clear_hightlight_color()
 	ppMap._update_map()
 	_set_text()
 
 
 func _set_text():
-	var dic = scoreMgr._get_score_data()
+	var dic = _g.scoreMgr._get_score_data()
 	ppLabelSummary.bbcode_text = dic["summary"]
 	ToolsMisc._RichLabelAdjustHeight(ppLabelSummary)
 	ppLabelDatetime.bbcode_text = dic["datetime"]
@@ -38,8 +23,8 @@ func _set_text():
 
 
 func _on_BtnBack_pressed():
-	dataMgr._save()
-	panelMgr.closePanel_animation(self)
+	_g.dataMgr._save()
+	_g.panelMgr.closePanel_animation(self)