+++ game/Scripts/Math/MathScoreOneTips.gd
@@ -1,14 +1,12 @@
 extends Label
 
-var logMgr
-var dataMgr
+var _g: Globals
 
 func _ready():
 	pass
 
 func _set_data(data: Array):
-	logMgr = get_node("/root/GLog")
-	dataMgr = get_node("/root/GData")
+	_g = get_node("/root/Globals")
 #	logMgr._debug("[MathScoreOneTips] set data:")
 #	logMgr._debug(data)
 	if data.size() == 3:
@@ -16,4 +14,4 @@ func _set_data(data: Array):
 	elif data.size() == 4:
 		text = str(data[0]) + " " + data[2] + " " + str(data[1]) + ": " + ("%.2f" % [data[3]]) + "s"
 		if data[3] <= 60:
-			dataMgr._addScore(data)
+			_g.dataMgr._addScore(data)