+++ game/Scripts/Misc/MiscChangeLog.gd
@@ -1,22 +1,10 @@
-extends Control
+extends PanelBasic
 
 onready var ppLabel = $VBoxContainer/Menu0/Label
 
-var panelMgr
-
-func _ready():
-	panelMgr = get_node("/root/PanelMgr")
-
-
-func _setRectSize(ssize: Vector2):
-	rect_position = Vector2(0, 0)
-	rect_size = ssize
-	$BasicBg.rect_min_size = ssize
-	$VBoxContainer.rect_min_size = ssize
-
 
 func _on_BtnBack_pressed():
-	panelMgr.closePanel_animation(self)
+	_g.panelMgr.closePanel_animation(self)
 
 
 func _on_ScrollContainer_page_changed(page):