+++ game/Scripts/Misc/MiscSetting.gd
@@ -1,31 +1,14 @@
-extends Control
+extends PanelBasic
 
-const PANEL_HUD_LAYER = -1
-const PANEL_PARTICLES_LAYER = 1
-const PANEL_NORMAL_LAYER = 3
-
-var dataMgr
-var panelMgr
-var logMgr
 
 onready var ppClearData : Button = $VBoxContainer/Menu1/HBoxContainer2/BtnClearData
+
 const CLEAR_DEFAULT_BTN_NAME = "ClearData"
 const CLEAR_RESET_TIME = 1
 const CLEAR_DEFAULT_STATE = 8
+
 var local_time: float = CLEAR_RESET_TIME
 var clearStage: int = CLEAR_DEFAULT_STATE
-
-func _ready():
-	panelMgr = get_node("/root/PanelMgr")
-	dataMgr = get_node("/root/GData")
-	logMgr = get_node("/root/GLog")
-
-
-func _setRectSize(ssize: Vector2):
-	rect_position = Vector2(0, 0)
-	rect_size = ssize
-	$BasicBg.rect_min_size = ssize
-	$VBoxContainer.rect_min_size = ssize
 
 
 func _process(delta):
@@ -42,16 +25,16 @@ func _resetBtnClear():
 
 
 func _on_BtnBack_pressed():
-	panelMgr.closePanel_animation(self)
+	_g.panelMgr.closePanel_animation(self)
 
 
 func _on_BtnLog_pressed():
-	if !panelMgr.openPanel("MiscTTLog", PANEL_PARTICLES_LAYER):
-		panelMgr.closePanel_name("MiscTTLog")
+	if !_g.panelMgr.openPanel("MiscTTLog", PanelMgr.PANEL_LAYER.PARTICLES_LAYER):
+		_g.panelMgr.closePanel_name("MiscTTLog")
 
 
 func _on_BtnPrintData_pressed():
-	dataMgr._printMetaData()
+	_g.dataMgr._printMetaData()
 
 
 func _on_BtnClearData_pressed():
@@ -60,14 +43,22 @@ func _on_BtnClearData_pressed():
 		clearStage -= 1
 		ppClearData.text = CLEAR_DEFAULT_BTN_NAME + "[" + str(clearStage) + "]"
 	else:
-		dataMgr._clearData()
+		_g.dataMgr._clearData()
 		yield(get_tree().create_timer(1), "timeout")
 		_resetBtnClear()
 
 
 func _on_BtnRefreshMathMap_pressed():
-	dataMgr._init_color_map()
-	dataMgr._save()
+	_g.dataMgr._init_color_map()
+	_g.dataMgr._save()
+
+
+func _on_BtnFPS_pressed():
+	if !_g.panelMgr.openPanel("MiscFPS", PanelMgr.PANEL_LAYER.HUD_LAYER):
+		_g.panelMgr.closePanel_name("MiscFPS")
+
+
+