+++ game/Scripts/Solar/SolarSystem.gd
@@ -1,0 +1,13 @@
+extends PanelBasic
+
+
+func _on_BtnBack_pressed():
+	_g.panelMgr.closePanel_animation(self)
+
+
+func _on_BtnPrevDay_pressed():
+	pass # Replace with function body.
+
+
+func _on_BtnNextDay_pressed():
+	pass # Replace with function body.