+++ game/Scripts/System/Globals.gd
@@ -1,10 +1,17 @@
 extends Node
 
-var logMgr
-var NinePadPositiveLayoutStyle: bool
+var logMgr: LogMgr
+var panelMgr: PanelMgr
+var dataMgr: GData
+var scoreMgr: GScore
 
 func _ready():
-	logMgr = get_node("/root/GLog")
+	logMgr = get_node("/root/LogMgr")
+	panelMgr = get_node("/root/PanelMgr")
+	dataMgr = get_node("/root/GData")
+	scoreMgr = get_node("/root/GScore")
+	
 	logMgr._log("[Globals] ready")
-	NinePadPositiveLayoutStyle = false
+	
+